Extract SBD and stopped resources to components
